# docs/openapi.front.yaml
openapi: 3.1.0
info:
  title: Lianchuan Storefront API
  version: 0.1.0
  summary: 前台电商网站接口契约
  description: |
    本文档定义前台商城所需的 REST API 契约，覆盖首页、分类、搜索、商品详情、购物车、结账、地址、订单、认证、会员资料、询价、收藏等能力。
    该文档是前台与服务端联调的唯一接口真源。
servers:
  - url: http://localhost:5100
    description: Local admin API (vexmotor-admin)
  - url: https://example.com
    description: Production placeholder
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        前台 JWT，由 POST /api/front/auth/login 或 POST /api/front/auth/register 返回的 token 字段；
        后续请求放在 Header Authorization Bearer token。
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
...
    LocaleHeader:
      name: x-vex-locale
      in: header
      description: 前台语言代码；与 site_languages 绑定，同时决定价格币种
      schema:
        type: string
        example: en
    LocaleQueryParam:
      name: locale
      in: query
      description: 可选语言覆盖（优先级低于 x-vex-locale）
      schema:
        type: string
        example: en
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type:
            - array
            - 'null'
          items:
            type: object
            additionalProperties: true
      required: [code, message]
      additionalProperties: false
...
    AcademyStat:
      type: object
      properties:
        label:
          type: string
        value:
          type: string
      required: [label, value]
    AcademyCertificateListItem:
      type: object
      properties:
        slug:
          type: string
        href:
          type: string
        title:
          type: string
        summary:
          type: string
        coverImage:
          type: string
          description: 公开封面 URL（R2 或预设图）
        teacherCount:
          type: integer
        studentCount:
          type: integer
        courseCount:
          type: integer
      required: [slug, href, title, summary, coverImage, teacherCount, studentCount, courseCount]
    AcademyCertificateCourseItem:
      type: object
      properties:
        slug:
          type: string
        href:
          type: string
        title:
          type: string
        summary:
          type: string
        coverImage:
          type: string
        sortOrder:
          type: integer
        units:
          type: array
          items:
            $ref: '#/components/schemas/AcademyCertificateUnitItem'
      required: [slug, href, title, summary, coverImage, sortOrder, units]
    AcademyCertificateUnitItem:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        sortOrder:
          type: integer
      required: [id, title, sortOrder]
    AcademyLessonItem:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        videoUrl:
          type: string
        durationSeconds:
          type: integer
        durationLabel:
          type: string
        sortOrder:
          type: integer
      required: [id, title, description, videoUrl, durationSeconds, durationLabel, sortOrder]
    AcademyUnitItem:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        coverImage:
          type: string
        sortOrder:
          type: integer
        lessons:
          type: array
          items:
            $ref: '#/components/schemas/AcademyLessonItem'
      required: [id, title, coverImage, sortOrder, lessons]
    AcademyCertificateListResponse:
      type: object
      properties:
        locale:
          type: string
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/AcademyCertificateListItem'
      required: [locale, page, pageSize, total, items]
    AcademyCertificateDetail:
      type: object
      properties:
        slug:
          type: string
        locale:
          type: string
        title:
          type: string
        summary:
          type: string
        description:
          type: string
          description: 富文本 HTML
        coverImage:
          type: string
        teacherCount:
          type: integer
        studentCount:
          type: integer
        stats:
          type: array
          items:
            $ref: '#/components/schemas/AcademyStat'
        learnings:
          type: array
          items:
            type: string
        skills:
          type: array
          items:
            type: string
        tools:
          type: array
          items:
            type: string
        courses:
          type: array
          items:
            $ref: '#/components/schemas/AcademyCertificateCourseItem'
        seo:
          type: object
          properties:
            title:
              type: string
            description:
              type: string
          required: [title, description]
      required: [slug, locale, title, summary, description, coverImage, teacherCount, studentCount, stats, learnings, skills, tools, courses, seo]
    AcademyCourseListItem:
      type: object
      properties:
        slug:
          type: string
        href:
          type: string
        title:
          type: string
        summary:
          type: string
        coverImage:
          type: string
        teacherCount:
          type: integer
        studentCount:
          type: integer
      required: [slug, href, title, summary, coverImage, teacherCount, studentCount]
    AcademyCourseListResponse:
      type: object
      properties:
        locale:
          type: string
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/AcademyCourseListItem'
      required: [locale, page, pageSize, total, items]
    AcademyCourseCertificateLink:
      type: object
      properties:
        slug:
          type: string
        href:
          type: string
        title:
          type: string
      required: [slug, href, title]
    AcademyCourseDetail:
      type: object
      properties:
        slug:
          type: string
        locale:
          type: string
        title:
          type: string
        summary:
          type: string
        description:
          type: string
        coverImage:
          type: string
        teacherCount:
          type: integer
        studentCount:
          type: integer
        stats:
          type: array
          items:
            $ref: '#/components/schemas/AcademyStat'
        learnings:
          type: array
          items:
            type: string
        skills:
          type: array
          items:
            type: string
        tools:
          type: array
          items:
            type: string
        units:
          type: array
          items:
            $ref: '#/components/schemas/AcademyUnitItem'
        certificates:
          type: array
          items:
            $ref: '#/components/schemas/AcademyCourseCertificateLink'
        seo:
          type: object
          properties:
            title:
              type: string
            description:
              type: string
          required: [title, description]
      required: [slug, locale, title, summary, description, coverImage, teacherCount, studentCount, stats, learnings, skills, tools, units, certificates, seo]
paths:
...
  /api/front/academy/certificates:
    get:
      tags: [Academy]
      summary: 获取已发布证书列表
      parameters:
        - $ref: '#/components/parameters/LocaleHeader'
        - $ref: '#/components/parameters/LocaleQueryParam'
        - $ref: '#/components/parameters/PageParam'
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 24
            maximum: 50
      responses:
        '200':
          description: 证书分页列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcademyCertificateListResponse'
  /api/front/academy/certificates/{slug}:
    get:
      tags: [Academy]
      summary: 获取证书详情（含有序课程摘要）
      parameters:
        - $ref: '#/components/parameters/LocaleHeader'
        - $ref: '#/components/parameters/LocaleQueryParam'
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 证书详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcademyCertificateDetail'
        '404':
          description: 未找到
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/front/academy/courses:
    get:
      tags: [Academy]
      summary: 获取已发布课程列表
      parameters:
        - $ref: '#/components/parameters/LocaleHeader'
        - $ref: '#/components/parameters/LocaleQueryParam'
        - $ref: '#/components/parameters/PageParam'
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 24
            maximum: 50
      responses:
        '200':
          description: 课程分页列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcademyCourseListResponse'
  /api/front/academy/courses/{slug}:
    get:
      tags: [Academy]
      summary: 获取课程详情（含所属证书链接）
      parameters:
        - $ref: '#/components/parameters/LocaleHeader'
        - $ref: '#/components/parameters/LocaleQueryParam'
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 课程详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcademyCourseDetail'
        '404':
          description: 未找到
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...